Split and cleanup relative stages
